hyperexplainer/backend/requirements.txt
Flask==2.2.5
google-generativeai>=0.3.0
google-cloud-storage==2.7.0
google-auth==2.20.0